requirements.txt
Sanic<0.8
Jinja2<2.11
walrus<0.6
requests<3
PyYAML
munch==2.2.0
